backend/requirements.txt
fastapi
uvicorn[standard]
pyyaml
git+https://github.com/nlesc-nano/CAT

